requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Data and preprocessing
numpy>=1.24.0
Pillow>=9.0.0
scikit-learn>=1.3.0

# Visualization and utilities
matplotlib>=3.7.0
tqdm>=4.65.0
umap-learn>=0.5.3

# Configuration
PyYAML>=6.0

# Download utilities
gdown>=4.7.0

# Optional: For distributed training
# torch-distributed-fp16-compress>=0.1.0
# horovod>=0.27.0
